requirements.txt
Flask==3.0.2
tinydb==4.8.0
python-dotenv==1.0.1